Use size_t for buffers representing data in memory rather than int64_t. (#2285)
* Use size_t for buffers representing data in memory rather than int64_t.

* add a cast from size_t to int64_t

* Address feedback - OnRead return size_t.

* Change Read and ReadToCount to return size_t instead of int64_t.

* Update curl usage and ReadFromSocket.

* Missed one signature.

* Update changelog.

* Update transport layer to use the right size_t size.

* Fix curl use of MBS.

* Cast to size_t since that's all memory stream supprts.

* Add static casts to size_t as temporary workarounds.

* update storage tests.

* More test fixes.

+++ sdk/core/azure-core/inc/azure/core/internal/io/null_body_stream.hpp
@@ -18,7 +18,7 @@ namespace Azure { namespace Core { namespace IO { namespace _internal {
    */
   class NullBodyStream final : public BodyStream {
   private:
-    int64_t OnRead(uint8_t* buffer, int64_t count, Azure::Core::Context const& context) override
+    size_t OnRead(uint8_t* buffer, size_t count, Azure::Core::Context const& context) override
     {
       (void)context;
       (void)buffer;
