+++ sdk/core/azure-core/test/ut/bodystream.cpp
@@ -26,7 +26,7 @@ using namespace Azure::Core;
 
 // Used to test virtual, default behavior of BodyStream.
 class TestBodyStream final : public BodyStream {
-  int64_t OnRead(uint8_t*, int64_t, Context const&) override { return 0; }
+  size_t OnRead(uint8_t*, size_t, Context const&) override { return 0; }
   int64_t Length() const override { return 0; }
 };
 
@@ -97,7 +97,7 @@ TEST(FileBodyStream, Read)
   // ReadToCount
   std::vector<uint8_t> buffer(FileSize * 2);
 
-  int64_t readSize = stream.ReadToCount(buffer.data(), 10);
+  size_t readSize = stream.ReadToCount(buffer.data(), 10);
   EXPECT_EQ(readSize, 10);
   EXPECT_EQ(buffer[10], 0);
